requirements.txt
boto3==1.9.251
botocore==1.12.251
docutils==0.15.2
jmespath==0.9.4
python-dateutil==2.8.0
python-dotenv==0.10.3
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.6
